Fixes and deprecation removal.
Added code to "fix" #206.  This attempts to get in front of the underlying issue, but even if it does, the code traps the error and generates a nicer message.

Removed some deprecated code listed in #205.  Some of this deprecation remains for compatibility.

Fixed the gradle-wrapper.properties to reference the right URL.