Refactor server config passing, and bug fixes.
For #208, the way the server configuration was passed around to the low-level API calls wasn't sufficient.  They also needed some properties from the client configuration (e.g. P4HOST).  A major refactoring of the code was necessary to allow these properties to be passed.

Another bug, a carry over from #209, resulted in loading configurations with wrong VCS root path mapping to the configuration.  A similar fix was necessary.

Cleaned up some of the logging while in here.