Partial fixes to #211, NPE fixes, and normalized path view.
Bug #211 can only be partially fixed, due to a problem with the IDE (see the bug report for details).

There were some additional NPEs discovered that came with the refactoring of the configuration passing, and with infrequently encoutnered logging.

The connection view pane now has the root paths shown with the right path separator character on Windows.